chore: consolidate decoded-data indexing into shared module
