+++ src/decoding/catchup/eth_calls.rs
@@ -22,6 +22,7 @@ use crate::decoding::eth_calls::{
     EthCallDecodingError, EventCallDecodeConfig,
 };
 use crate::decoding::types::{EthCallResult, EventCallResult};
+use crate::storage::decoded_index::scan_existing_decoded_files;
 use crate::transformations::DecodedCallsMessage;
 use crate::types::config::raw_data::RawDataCollectionConfig;
 
@@ -530,29 +531,6 @@ pub async fn catchup_decode_eth_calls(
     );
 
     Ok(())
-}
-
-/// Scan existing decoded files
-fn scan_existing_decoded_files(output_base: &Path) -> HashSet<String> {
-    let mut files = HashSet::new();
-
-    fn scan_recursive(dir: &Path, base: &Path, files: &mut HashSet<String>) {
-        if let Ok(entries) = std::fs::read_dir(dir) {
-            for entry in entries.flatten() {
-                let path = entry.path();
-                if path.is_dir() {
-                    scan_recursive(&path, base, files);
-                } else if path.extension().map(|e| e == "parquet").unwrap_or(false) {
-                    if let Ok(rel) = path.strip_prefix(base) {
-                        files.insert(rel.to_string_lossy().to_string());
-                    }
-                }
-            }
-        }
-    }
-
-    scan_recursive(output_base, output_base, &mut files);
-    files
 }
 
 /// Read regular call results from parquet
